+++ Cargo.toml
@@ -26,4 +26,6 @@ entlib-native-sha2 =                { path = "./crypto/sha2",               vers
 entlib-native-sha3 =                { path = "./crypto/sha3",               version = "1.1.2-Alpha1" }
 entlib-native-chacha20 =            { path = "./crypto/chacha20",           version = "1.1.2-Alpha1" }
 entlib-native-key-establishment =   { path = "./crypto/key-establishment",  version = "1.1.2-Alpha1" }
+entlib-native-digital-signature =   { path = "./crypto/digital-signature",  version = "1.1.2-Alpha1" }
 entlib-native-tls =                 { path = "./crypto/tls",                version = "1.1.2-Alpha1" }
+entlib-native-hkdf =                { path = "./crypto/hkdf",               version = "1.1.2-Alpha1" }
