+++ crypto/digital-signature/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod ed25519;
+pub mod ed448;
