+++ crypto/hkdf/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "entlib-native-hkdf"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+
+[dependencies]
